加群申请，阿里巴巴规范

+++ src/main/java/com/example/im/entity/Login.java
@@ -19,7 +19,7 @@ public class Login {
     @Id
     private String id;
     private String ip;
-    private Byte port;
+    private Integer port;
     private Boolean status;
     private String userId;
 }
